changed sfcparse to a new created package strucutre
