+++ src/sfcparse/__xml/xmlbuildmanual.py
@@ -1,0 +1,18 @@
+# xmlbuildmanual
+#########################################################################################################
+# Imports
+import xml.etree.ElementTree as __xml_etree
+
+
+#########################################################################################################
+# Build manual xml data
+def xmlbuildmanual() -> __xml_etree:
+    """
+    Returns a empty xml ElementTree obj to build/work with xml data
+    
+    Assign the output to var
+
+    This is using the native xml library via etree shipped with the python standard libray.
+    For more information on the xml.etree api, visit: https://docs.python.org/3/library/xml.etree.elementtree.html#module-xml.etree.ElementTree
+    """
+    return __xml_etree
